Add auth functionality and extract from module.
