+++ .vscode/launch.json
@@ -11,6 +11,7 @@
             "url": "http://localhost:4200",
             "port": 9222,
             "webRoot": "${workspaceFolder}",
+            "userDataDir": "C:\\Users\\William\\Programming\\chrome-dev",
             "runtimeExecutable": "C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe",
             "runtimeArgs": [
                 "--runtime-debugger-port=9222"
@@ -23,6 +24,7 @@
             "url": "http://localhost:4200",
             "webRoot": "${workspaceFolder}",
             "preLaunchTask": "debug",
+            "userDataDir": "C:\\Users\\William\\Programming\\chrome-dev",
             "runtimeExecutable": "C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe",
             "runtimeArgs": [
                 "--runtime-debugger-port=9222"
