--- src/app/firebase-auth/firebase-auth.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { FirebaseAuthService } from './firebase-auth.service';
-
-describe('FirebaseAuthService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [FirebaseAuthService]
-    });
-  });
-
-  it('should be created', inject([FirebaseAuthService], (service: FirebaseAuthService) => {
-    expect(service).toBeTruthy();
-  }));
-});
